fix: 🐛 Manually determine and intercept the Api request
